Moved test classes to right folders and removed benchmark code (is on other repo)
